% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdid.R
\name{sdid_att}
\alias{sdid_att}
\title{High-level SDID across cohorts in an unbalanced panel}
\usage{
sdid_att(
  id,
  t,
  y,
  g,
  L = 4L,
  F = 4L,
  min_cov = 0.8,
  lambda_unit = 0.001,
  lambda_time = 0.001,
  maxit = 5000L,
  tol = 1e-08
)
}
\description{
High-level SDID across cohorts in an unbalanced panel
}
